requirements.txt
dj_rest_auth==3.0.0
Django==4.1.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
rest_framework_simplejwt==0.0.2
